@@ -2,13 +2,23 @@
 
 public class TestComponent : UiComponent
 {
-    private readonly Random rand = new Random();
-
+    private readonly Random rand = new();
+    private ColorDefinition _color;
+    
+    public TestComponent()
+    {
+        _color = GetRandomColor();
+    }
+    
     public override Div Render()
     {
         return new Div
         {
-            new Div().Width(50, SizeKind.Percentage).Height(50, SizeKind.Percentage).Color(GetRandomColor())
+            new Div().Width(50, SizeKind.Percentage).Height(50, SizeKind.Percentage).Color(_color).OnClick(
+                () =>
+                {
+                    _color = GetRandomColor();
+                })
         }.MAlign(MAlign.Center).XAlign(XAlign.Center);
     }
 
